Fix wrong type in audio stats
Field mrsAudioReceiverStats::total_samples_received had invalid type
double instead of uint64_t. Fix type, thereby silencing compiler warning
about implicit cast.

+++ libs/Microsoft.MixedReality.WebRTC.Native/src/interop/interop_api.h
@@ -757,7 +757,7 @@ struct mrsAudioReceiverStats {
   const char* track_identifier;
   double audio_level;
   double total_audio_energy;
-  double total_samples_received;
+  uint64_t total_samples_received;
   double total_samples_duration;
 
   int64_t rtp_stats_timestamp_us;
